added --statistics flag

+++ inkcpp/array.h
@@ -29,7 +29,7 @@ public:
 		}
 	}
 
-	config::statistics::container statistics() const { return {_capacity, _size}; }
+	config::statistics::container statistics() const { return {static_cast<int>(_capacity), static_cast<int>(_size)}; }
 
 	virtual ~managed_array()
 	{
