+++ inkcpp/simple_restorable_stack.h
@@ -29,7 +29,7 @@ public:
 	virtual ~simple_restorable_stack() = default;
 
 	config::statistics::container statistics() const {
-		return {_size, _pos};
+		return {static_cast<int>(_size), static_cast<int>(_pos)};
 	}
 
 	void     push(const T& value);
